requirements.txt
langchain
sentence-transformers
psycopg2
neo4j
fastapi
spacy
streamlit
prometheus-client
grafana-client
mlflow
ragas
docker
python-dotenv
uvicorn
